wip(today's fortune): Santa Claus is watching!
